Changes to delegate authentication to an external service
There are two changes here:

- Return the auth server's URL from the .well-known/oauth-protected-resource endpoint
- Ask the auth server to check the client's tokens to make sure they are valid (RFC7662)

The official TypeScript client at https://www.npmjs.com/package/@atproto/oauth-client insists on
using "DPoP" as the token type. This might mean that millipds should somehow share its DPoP state
with the auth server, too, and check the DPoP header (@dpop_protected) on requests to any protected
endpoint, not just the PAR endpoint.

The configuration is provided via environment variables; it probably belongs in the "config" table
instead, but I couldn't see how this project manages database schema changes, so I've done it the
easy way.

Of course millipds and the auth server will need to have the same idea of which accounts exist.
Keeping the two in sync will need to be the responsibility of some external service.